+++ POS/View/frmUrunlerView.cs
@@ -62,13 +62,17 @@ namespace POS.View
             }
             if (guna2DataGridView1.CurrentCell.OwningColumn.Name == "dgvdel")
             {
-                int id = Convert.ToInt32(guna2DataGridView1.CurrentRow.Cells["dgvid"].Value);
-                string qry = "Delete from Urunler where uID= " + id + "";
-                Hashtable ht = new Hashtable();
-                MainClass.SQL(qry, ht);
+                if (MessageBox.Show("Silmek istediğinizden Emin Misiniz?", "Ürün Silme", MessageBoxButtons.YesNo, MessageBoxIcon.Question) == DialogResult.Yes) 
+                {
+                    int id = Convert.ToInt32(guna2DataGridView1.CurrentRow.Cells["dgvid"].Value);
+                    string qry = "Delete from Urunler where uID= " + id + "";
+                    Hashtable ht = new Hashtable();
+                    MainClass.SQL(qry, ht);
 
-                MessageBox.Show("başarıyla Silindi");
-                GetData();
+                    MessageBox.Show("başarıyla Silindi");
+                    GetData();
+                }
+                   
             }
 
         }
